docker-compose, Kafka Broker, data generation and publishing in docker container

+++ app/generator.py
@@ -1,7 +1,10 @@
 from helper.file_downloader import download_all_files
 from helper.zip_extractor import extract_all_product_txt_files
+from helper.kafka_publisher import publish_staged
 import shutil
 import os
+from kafka.admin import KafkaAdminClient, NewTopic
+from kafka import KafkaClient
 
 # Download URL of the DWDs for hourly wind data
 DOWNLOAD_URL = "https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/wind/recent/"
@@ -10,8 +13,12 @@ DOWNLOAD_FOLDER = os.path.join("tmp", "input")
 # File in the weather station zip file that contains the wind data. The rest only has metadata
 EXTRACT_REGEX = r"^produkt_ff_stunde_\w*\.txt$"
 STAGED_FOLDER = os.path.join("tmp", "staged")
+TOPIC_REGEX = r"_(\d+)\.txt$"
 
 DESCRIPTION_FILE = os.path.join(DOWNLOAD_FOLDER, "FF_Stundenwerte_Beschreibung_Stationen.txt")
+
+KAFKA_BROKER = os.environ.get("KAFKA_BROKER")
+print(f"Kafka Broker: {KAFKA_BROKER}")
 
 def run():
     """
@@ -26,8 +33,11 @@ def run():
     if os.path.exists(DESCRIPTION_FILE):
         shutil.copy2(DESCRIPTION_FILE, STAGED_FOLDER)
     
+    publish_staged(STAGED_FOLDER, KAFKA_BROKER, TOPIC_REGEX)
+    
     # Remove the temporary zip download folder
-    shutil.rmtree(DOWNLOAD_FOLDER)
+    if os.path.exists(DOWNLOAD_FOLDER):
+        shutil.rmtree(DOWNLOAD_FOLDER)
 
 
 if __name__ == "__main__":
